analytics: Extract test_stats_views.
This will help make the analytics views codebase more readable.
